Refactor some code to make the climate better
